+++ src/Twilio/Rest/Preview/TrustedComms/Business/InsightsPage.php
@@ -7,11 +7,14 @@
  * /       /
  */
 
-namespace Twilio\Rest\Verify\V2\Service;
+namespace Twilio\Rest\Preview\TrustedComms\Business;
 
 use Twilio\Page;
 
-class RateLimitPage extends Page {
+/**
+ * PLEASE NOTE that this class contains preview products that are subject to change. Use them with caution. If you currently do not have developer preview access, please contact help@twilio.com.
+ */
+class InsightsPage extends Page {
     public function __construct($version, $response, $solution) {
         parent::__construct($version, $response);
 
@@ -20,7 +23,7 @@ class RateLimitPage extends Page {
     }
 
     public function buildInstance(array $payload) {
-        return new RateLimitInstance($this->version, $payload, $this->solution['serviceSid']);
+        return new InsightsInstance($this->version, $payload, $this->solution['businessSid']);
     }
 
     /**
@@ -29,6 +32,6 @@ class RateLimitPage extends Page {
      * @return string Machine friendly representation
      */
     public function __toString() {
-        return '[Twilio.Verify.V2.RateLimitPage]';
+        return '[Twilio.Preview.TrustedComms.InsightsPage]';
     }
 }