+++ src/Twilio/Rest/Verify/V2/Service/RateLimit/BucketPage.php
@@ -11,9 +11,6 @@ namespace Twilio\Rest\Verify\V2\Service\RateLimit;
 
 use Twilio\Page;
 
-/**
- * PLEASE NOTE that this class contains beta products that are subject to change. Use them with caution.
- */
 class BucketPage extends Page {
     public function __construct($version, $response, $solution) {
         parent::__construct($version, $response);