+++ src/Twilio/Rest/Verify/V2/Service/RateLimitContext.php
@@ -17,8 +17,6 @@ use Twilio\Values;
 use Twilio\Version;
 
 /**
- * PLEASE NOTE that this class contains beta products that are subject to change. Use them with caution.
- *
  * @property \Twilio\Rest\Verify\V2\Service\RateLimit\BucketList $buckets
  * @method \Twilio\Rest\Verify\V2\Service\RateLimit\BucketContext buckets(string $sid)
  */