+++ src/Twilio/Rest/Verify/V2/Service/RateLimitOptions.php
@@ -12,9 +12,6 @@ namespace Twilio\Rest\Verify\V2\Service;
 use Twilio\Options;
 use Twilio\Values;
 
-/**
- * PLEASE NOTE that this class contains beta products that are subject to change. Use them with caution.
- */
 abstract class RateLimitOptions {
     /**
      * @param string $description Description of this Rate Limit