+++ tests/Twilio/Integration/Lookups/V1/PhoneNumberTest.php
@@ -35,37 +35,14 @@ class PhoneNumberTest extends HolodeckTestCase {
             200,
             '
             {
-                "caller_name": {
-                    "caller_name": "Delicious Cheese Cake",
-                    "caller_type": "CONSUMER",
-                    "error_code": null
-                },
-                "carrier": {
-                    "error_code": null,
-                    "mobile_country_code": "310",
-                    "mobile_network_code": "456",
-                    "name": "verizon",
-                    "type": "mobile"
-                },
-                "fraud": {
-                    "error_code": null,
-                    "mobile_country_code": "310",
-                    "mobile_network_code": "456",
-                    "advanced_line_type": "voip",
-                    "caller_name": "Delicious Cheese Cake",
-                    "is_ported": false,
-                    "last_ported_date": "2018-05-01 04:05:11"
-                },
+                "caller_name": null,
+                "carrier": null,
+                "fraud": null,
+                "add_ons": null,
                 "country_code": "US",
                 "national_format": "(510) 867-5310",
                 "phone_number": "+15108675310",
-                "add_ons": {
-                    "status": "successful",
-                    "message": null,
-                    "code": null,
-                    "results": {}
-                },
-                "url": "https://lookups.twilio.com/v1/PhoneNumbers/phone_number"
+                "url": "https://lookups.twilio.com/v1/PhoneNumbers/+15108675310"
             }
             '
         ));
@@ -93,7 +70,35 @@ class PhoneNumberTest extends HolodeckTestCase {
                 "phone_number": "+15108675310",
                 "fraud": null,
                 "add_ons": null,
-                "url": "https://lookups.twilio.com/v1/PhoneNumbers/phone_number"
+                "url": "https://lookups.twilio.com/v1/PhoneNumbers/+15108675310?Type=carrier"
+            }
+            '
+        ));
+
+        $actual = $this->twilio->lookups->v1->phoneNumbers("+15017122661")->fetch();
+
+        $this->assertNotNull($actual);
+    }
+
+    public function testFetchCarrierInternationalResponse() {
+        $this->holodeck->mock(new Response(
+            200,
+            '
+            {
+                "caller_name": null,
+                "carrier": {
+                    "error_code": null,
+                    "mobile_country_code": null,
+                    "mobile_network_code": null,
+                    "name": "Vodafone Business Solutions",
+                    "type": "landline"
+                },
+                "country_code": "GB",
+                "national_format": "020 7765 1182",
+                "phone_number": "+4402077651182",
+                "fraud": null,
+                "add_ons": null,
+                "url": "https://lookups.twilio.com/v1/PhoneNumbers/+4402077651182?Type=carrier"
             }
             '
         ));
@@ -119,7 +124,7 @@ class PhoneNumberTest extends HolodeckTestCase {
                 "national_format": "(510) 867-5310",
                 "phone_number": "+15108675310",
                 "add_ons": null,
-                "url": "https://lookups.twilio.com/v1/PhoneNumbers/phone_number"
+                "url": "https://lookups.twilio.com/v1/PhoneNumbers/+15108675310?Type=caller-name"
             }
             '
         ));
@@ -156,7 +161,7 @@ class PhoneNumberTest extends HolodeckTestCase {
                     "code": null,
                     "results": {}
                 },
-                "url": "https://lookups.twilio.com/v1/PhoneNumbers/+15108675310"
+                "url": "https://lookups.twilio.com/v1/PhoneNumbers/+15108675310?Type=carrier&Type=caller-name"
             }
             '
         ));
@@ -231,7 +236,7 @@ class PhoneNumberTest extends HolodeckTestCase {
                         }
                     }
                 },
-                "url": "https://lookups.twilio.com/v1/PhoneNumbers/+12127363100?Type=carrier"
+                "url": "https://lookups.twilio.com/v1/PhoneNumbers/+12127363100?Type=caller-name"
             }
             '
         ));