+++ tests/Twilio/Integration/Wireless/V1/RatePlanTest.php
@@ -91,6 +91,7 @@ class RatePlanTest extends HolodeckTestCase {
                         "sid": "WPaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                         "usage_notification_method": "POST",
                         "usage_notification_url": "https://callback.com",
+                        "data_limit_strategy": "block",
                         "url": "https://wireless.twilio.com/v1/RatePlans/WPaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
                     }
                 ]
@@ -143,6 +144,7 @@ class RatePlanTest extends HolodeckTestCase {
                 "sid": "WPaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "usage_notification_method": "POST",
                 "usage_notification_url": "https://callback.com",
+                "data_limit_strategy": "block",
                 "url": "https://wireless.twilio.com/v1/RatePlans/WPaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
             }
             '
@@ -176,6 +178,7 @@ class RatePlanTest extends HolodeckTestCase {
                 "unique_name": "unique_name",
                 "data_enabled": true,
                 "data_limit": 1000,
+                "data_limit_strategy": "block",
                 "data_metering": "payg",
                 "date_created": "2019-07-30T20:00:00Z",
                 "date_updated": "2019-07-30T20:00:00Z",
@@ -243,6 +246,7 @@ class RatePlanTest extends HolodeckTestCase {
                 "sid": "WPaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "usage_notification_method": "POST",
                 "usage_notification_url": "https://callback.com",
+                "data_limit_strategy": "block",
                 "url": "https://wireless.twilio.com/v1/RatePlans/WPaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
             }
             '