+++ CHANGES.md
@@ -1,6 +1,21 @@
 twilio-php Changelog
 ====================
 
+[2019-12-18] Version 5.41.1
+---------------------------
+**Preview**
+- Add `/Insights/SuccessRate` endpoint for Businesses Branded Calls (Verified by Twilio)
+
+**Studio**
+- StudioV2 API in beta
+
+**Verify**
+- Add `MailerSid` property to Verify Service resource.
+
+**Wireless**
+- Added `data_limit_strategy` to Rate Plan resource.
+
+
 [2019-12-12] Version 5.41.0
 ---------------------------
 **Api**